backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.12.0
httpx>=0.26.0
python-dotenv>=1.0.0
prefect>=2.14.0
moviepy>=1.0.3
elevenlabs>=1.0.0
google-api-python-client>=2.111.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
Pillow>=10.0.0
numpy>=1.24.0
